+++ src/Phan/Language/Element/ClassConstant.php
@@ -65,6 +65,37 @@ class ClassConstant extends ClassElement implements ConstantInterface
     }
 
     /**
+     * Create an alias from a trait use, which is treated as though it was defined in $clazz
+     * E.g. if you import a trait's class constant as private/protected, it becomes private/protected **to the class which used the trait**
+     *
+     * The resulting alias doesn't inherit the Node of the method body, so aliases won't have a redundant analysis step.
+     *
+     * @param Clazz $clazz - The class to treat as the defining class of the alias. (i.e. the inheriting class)
+     */
+    public function createUseAlias(Clazz $clazz): ClassConstant
+    {
+        $constant_fqsen = FullyQualifiedClassConstantName::make(
+            $clazz->getFQSEN(),
+            $this->name
+        );
+
+        $constant = new ClassConstant(
+            $this->getContext(),
+            $this->name,
+            $this->getUnionType(),
+            $this->getFlags(),
+            $constant_fqsen
+        );
+        $constant->setPhanFlags($this->getPhanFlags() & ~(Flags::IS_OVERRIDE | Flags::IS_OVERRIDDEN_BY_ANOTHER));
+
+        $defining_fqsen = $this->getDefiningFQSEN();
+        if ($constant->isPublic()) {
+            $constant->setDefiningFQSEN($defining_fqsen);
+        }
+        return $constant;
+    }
+
+    /**
      * Override the default getter to fill in a future
      * union type if available.
      */