+++ NEWS.md
@@ -4,7 +4,10 @@ Phan NEWS
 -----------------------
 
 New Features(Analysis):
-- Support parsing php 8.2's disjunctive normal form types (e.g. `A|(B&C)` (https://wiki.php.net/rfc/dnf_types).
+- Support parsing php 8.2's disjunctive normal form types (e.g. `A|(B&C)` (https://wiki.php.net/rfc/dnf_types). (#4699)
+- Support php 8.2 class constants on trait RFC. (#4687)
+  Emit `PhanCompatibleTraitConstant` when using constants on traits with a `minimum_target_php_version` below `'8.2'`
+  Emit `PhanAccessClassConstantOfTraitDirectly` when directly referencing the class constant on the trait declaration.
 
 Aug 08 2022, Phan 5.4.0
 -----------------------